passed validation logic to Validation.cs
